Move addNote function from noteModal to noteUtils
